backend/requirements.txt
flask
flask-cors
torch
transformers
accelerate
kafka-python
redis
spacy
gunicorn
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.7.0/ru_core_news_sm-3.7.0-py3-none-any.whl